fixed printf typing error

+++ Bacteria.cpp
@@ -15,7 +15,7 @@ Bacteria::Bacteria()
 	bacteria.push_back( new Bacterium() );
 	bacteria.push_back( new Bacterium() );
 	bacteria.push_back( new Bacterium() );
-	printf("Number of bacteria: %d\n", bacteria.size());
+	printf("Number of bacteria: %d\n", (int)bacteria.size());
 
 }
